+++ models/1_source/1_FPL_ALL_SEASONS.sql
@@ -5,7 +5,7 @@ WITH FPL_SEASON_2017 AS(
     CAST(Goals_Conceded AS INT64),
     CAST(Own_Goals AS INT64),
     CAST(SAVES AS INT64),
-    CAST(COST AS FLOAT),
+    CAST(COST AS float64),
     CAST(Points AS INT64),
     2017 AS SEASON_START
   FROM {{ source('fpl_historic', '2017') }}),
@@ -17,7 +17,7 @@ FPL_SEASON_2018 AS(
     CAST(Goals_Conceded AS INT64),
     CAST(Own_Goals AS INT64),
     CAST(SAVES AS INT64),
-    CAST(COST AS FLOAT),
+    CAST(COST AS float64),
     CAST(Points AS INT64),
     2018 AS SEASON_START
   FROM {{ source('fpl_historic', '2018') }}),
@@ -29,7 +29,7 @@ FPL_SEASON_2019 AS(
     CAST(Goals_Conceded AS INT64),
     CAST(Own_Goals AS INT64),
     CAST(SAVES AS INT64),
-    CAST(COST AS FLOAT),
+    CAST(COST AS float64),
     CAST(Points AS INT64),
     2019 AS SEASON_START
   FROM {{ source('fpl_historic', '2019') }}),
@@ -41,7 +41,7 @@ FPL_SEASON_2020 AS(
     CAST(Goals_Conceded AS INT64),
     CAST(Own_Goals AS INT64),
     CAST(SAVES AS INT64),
-    CAST(COST AS FLOAT),
+    CAST(COST AS float64),
     CAST(Points AS INT64),
     2020 AS SEASON_START
   FROM {{ source('fpl_historic', '2020') }}),
@@ -53,7 +53,7 @@ FPL_SEASON_2021 AS(
     CAST(Goals_Conceded AS INT64),
     CAST(Own_Goals AS INT64),
     CAST(SAVES AS INT64),
-    CAST(COST AS FLOAT),
+    CAST(COST AS float64),
     CAST(Points AS INT64),
     2021 AS SEASON_START
   FROM {{ source('fpl_historic', '2021') }}),
@@ -65,7 +65,7 @@ FPL_SEASON_2022 AS(
     CAST(Goals_Conceded AS INT64),
     CAST(Own_Goals AS INT64),
     CAST(SAVES AS INT64),
-    CAST(COST AS FLOAT),
+    CAST(COST AS float64),
     CAST(Points AS INT64),
     2022 AS SEASON_START
   FROM {{ source('fpl_historic', '2022') }}),
@@ -77,7 +77,7 @@ FPL_SEASON_2023 AS(
     CAST(Goals_Conceded AS INT64),
     CAST(Own_Goals AS INT64),
     CAST(SAVES AS INT64),
-    CAST(COST AS FLOAT),
+    CAST(COST AS float64),
     CAST(Points AS INT64),
     2023 AS SEASON_START
   FROM {{ source('fpl_historic', '2023') }})
